feat: Moved Chromosome class to a separate file
